♻️ refactor: refactor the layout provider
