+++ src/app/layout.tsx
@@ -1,50 +1,35 @@
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import { ResolvingViewport } from 'next';
+import { SessionProvider } from 'next-auth/react';
 import { cookies } from 'next/headers';
 import { PropsWithChildren } from 'react';
 import { isRtlLang } from 'rtl-detect';
+import GlobalProvider from 'src/layout/GlobalProvider';
 
 import Analytics from '@/components/Analytics';
 import { getServerConfig } from '@/config/server';
 import { DEFAULT_LANG, LOBE_LOCALE_COOKIE } from '@/const/locale';
-import {
-  LOBE_THEME_APPEARANCE,
-  LOBE_THEME_NEUTRAL_COLOR,
-  LOBE_THEME_PRIMARY_COLOR,
-} from '@/const/theme';
-import Layout from '@/layout/GlobalLayout';
-import { getAntdLocale } from '@/utils/locale';
+import { API_ENDPOINTS } from '@/services/_url';
 import { isMobileDevice } from '@/utils/responsive';
 
-import StyleRegistry from './StyleRegistry';
-
-const { ENABLE_OAUTH_SSO } = getServerConfig();
+const { ENABLE_OAUTH_SSO = false } = getServerConfig();
 
 const RootLayout = async ({ children }: PropsWithChildren) => {
-  // get default theme config to use with ssr
   const cookieStore = cookies();
-  const appearance = cookieStore.get(LOBE_THEME_APPEARANCE);
-  const neutralColor = cookieStore.get(LOBE_THEME_NEUTRAL_COLOR);
-  const primaryColor = cookieStore.get(LOBE_THEME_PRIMARY_COLOR);
+
   const lang = cookieStore.get(LOBE_LOCALE_COOKIE);
   const direction = isRtlLang(lang?.value || DEFAULT_LANG) ? 'rtl' : 'ltr';
 
-  const antdLocale = await getAntdLocale(lang?.value);
+  const content = ENABLE_OAUTH_SSO ? (
+    <SessionProvider basePath={API_ENDPOINTS.oauth}>{children}</SessionProvider>
+  ) : (
+    children
+  );
+
   return (
     <html dir={direction} lang={lang?.value || DEFAULT_LANG} suppressHydrationWarning>
       <body>
-        <StyleRegistry>
-          <Layout
-            antdLocale={antdLocale}
-            defaultAppearance={appearance?.value}
-            defaultLang={lang?.value}
-            defaultNeutralColor={neutralColor?.value as any}
-            defaultPrimaryColor={primaryColor?.value as any}
-            enableOAuthSSO={ENABLE_OAUTH_SSO}
-          >
-            {children}
-          </Layout>
-        </StyleRegistry>
+        <GlobalProvider>{content}</GlobalProvider>
         <Analytics />
         <SpeedInsights />
       </body>
