+++ src/features/DebugUI/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Icon } from '@lobehub/ui';
 import { App, FloatButton, Spin } from 'antd';
 import { DatabaseIcon, Loader2 } from 'lucide-react';
